namespace housekeeping for folder splits
