+++ src/Dhgms.GripeWithRoslyn.Analyzer/Analyzers/Language/BooleanTryMethodShouldBeUsedInLogicalNotIfStatementAnalyzer.cs
@@ -10,7 +10,7 @@ using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Diagnostics;
 
-namespace Dhgms.GripeWithRoslyn.Analyzer.Analyzers
+namespace Dhgms.GripeWithRoslyn.Analyzer.Analyzers.Language
 {
     /// <summary>
     /// Analyzer to check if TryParse is used in a logical if statement.
