+++ src/Dhgms.GripeWithRoslyn.Analyzer/Analyzers/Language/UseTypeofInsteadOfTypeGetTypeAnalyzer.cs
@@ -2,15 +2,11 @@
 // This file is licensed to you under the MIT license.
 // See the LICENSE file in the project root for full license information.
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using Dhgms.GripeWithRoslyn.Analyzer.Analyzers.Abstractions;
 using Dhgms.GripeWithRoslyn.Analyzer.CodeCracker.Extensions;
 using Microsoft.CodeAnalysis;
 
-namespace Dhgms.GripeWithRoslyn.Analyzer.Analyzers
+namespace Dhgms.GripeWithRoslyn.Analyzer.Analyzers.Language
 {
     /// <summary>
     /// Analyzer to suggest using typeof() instead of <see cref="M:System.Type.GetType"/>.
