+++ src/Dhgms.GripeWithRoslyn.Analyzer/Analyzers/MediatR/NameOfRequestHandlerShouldEndWithCommandHandlerOrQueryHandlerAnalyzer.cs
@@ -2,6 +2,7 @@
 // This file is licensed to you under the MIT license.
 // See the LICENSE file in the project root for full license information.
 
+using Dhgms.GripeWithRoslyn.Analyzer.Analyzers.Abstractions;
 using Dhgms.GripeWithRoslyn.Analyzer.CodeCracker.Extensions;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Diagnostics;
