+++ src/Dhgms.GripeWithRoslyn.Analyzer/Analyzers/ReactiveUi/ViewModelClassShouldInheritFromViewModelInterfaceAnalyzer.cs
@@ -5,7 +5,6 @@
 using System;
 using System.Collections.Immutable;
 using Dhgms.GripeWithRoslyn.Analyzer.CodeCracker.Extensions;
-using Dhgms.GripeWithRoslyn.UnitTests.Helpers;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
