+++ src/Dhgms.GripeWithRoslyn.UnitTests/Analyzers/UseDateTimeUtcNowInsteadofNowAnalyzertest.cs
@@ -4,6 +4,7 @@
 
 using Dhgms.GripeWithRoslyn.Analyzer;
 using Dhgms.GripeWithRoslyn.Analyzer.Analyzers;
+using Dhgms.GripeWithRoslyn.Analyzer.Analyzers.Runtime;
 using Dhgms.GripeWithRoslyn.UnitTests.Helpers;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Diagnostics;
