+++ src/Dhgms.GripeWithRoslyn.UnitTests/Analyzers/UseSystemTextJsonInsteadOfNewtonsoftJsonAnalyzerTests.cs
@@ -5,6 +5,7 @@
 using System.Reflection;
 using Dhgms.GripeWithRoslyn.Analyzer;
 using Dhgms.GripeWithRoslyn.Analyzer.Analyzers;
+using Dhgms.GripeWithRoslyn.Analyzer.Analyzers.Runtime;
 using Dhgms.GripeWithRoslyn.UnitTests.Helpers;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Diagnostics;
